More attempts at breaking source
git-svn-id: file:///svn/phpbb/trunk@4072 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/includes/ucp/ucp_register.php
@@ -81,7 +81,7 @@ class ucp_register extends ucp
 					'email'				=> '7,60', 
 					'email_confirm'		=> '7,60',
 				),
-				'int'	=> array('tz')
+				'float'	=> array('tz')
 			);
 			$data = $this->normalise_data($_POST, $normalise);
